cleanup: remove sitemesh from package/test names
