+++ grails-test-examples/app1/grails-app/controllers/functionaltests/layout/LayoutByConventionController.groovy
@@ -17,12 +17,9 @@
  *  under the License.
  */
 
-package functionaltests.sitemesh
+package functionaltests.layout
 
+class LayoutByConventionController {
 
-class LayoutTemplateController {
-
-    def index() { 
-        render template: 'partial'
-    }
+    def index() {}
 }
