+++ grails-test-examples/app1/grails-app/controllers/functionaltests/layout/LayoutSpecifiedByPropertyController.groovy
@@ -17,7 +17,7 @@
  *  under the License.
  */
 
-package functionaltests.sitemesh
+package functionaltests.layout
 
 class LayoutSpecifiedByPropertyController {
